backend/requirements.txt
Django>=4.2,<5.0
django-cors-headers>=4.0
djangorestframework>=3.14
python-dotenv>=1.0
mysqlclient>=2.2
django-filter>=23.0